Add SiaException for when the node isn't running, improve progress bar appearance on wallet page, improve rounding and displaying of BigDecimals

+++ app/src/main/java/com/vandyke/sia/data/models/wallet/TransactionData.kt
@@ -21,7 +21,7 @@ data class TransactionData @JsonCreator constructor(
         @JsonProperty(value = "transactionid")
         val transactionId: String,
         @JsonProperty(value = "confirmationheight")
-        val confirmationHeight: Int,
+        val confirmationHeight: BigDecimal,
         @JsonProperty(value = "confirmationtimestamp")
         val confirmationTimestamp: BigDecimal,
         @Ignore
@@ -33,7 +33,7 @@ data class TransactionData @JsonCreator constructor(
 
 
     constructor(transactionId: String,
-                confirmationHeight: Int,
+                confirmationHeight: BigDecimal,
                 confirmationTimestamp: BigDecimal) :
             this(transactionId, confirmationHeight, confirmationTimestamp, null, null)
